<feat, style, refactor>(exception, record, repository): 상품 유무 예외처리 추가, record 클래스 가독성 좋게 변경, containsKey로직 수정
ExceptionResponseDto : 예외 응답 반환 클래스
ProductNotFoundException : 상품 존재 x 예외
GlobalExceptionHandler : Controller의 예외 처리를 수행하는 클래스
containsKey : 쿼리문 앞에 !연산자를 제거해서 해당 id를 가진 상품이 있을 때 true를 반환하게 함

+++ src/main/java/gift/service/ProductService.java
@@ -4,6 +4,7 @@ import gift.dto.request.ProductRequestDto;
 import gift.dto.request.ProductUpdateRequestDto;
 import gift.dto.response.ProductResponseDto;
 import gift.entity.Product;
+import gift.exception.ProductNotFoundException;
 import gift.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +20,7 @@ public class ProductService implements ProductServiceInterface {
 
     //상품 단 건 조회
     public Product getProduct(long productId) {
+        if(!containsProduct(productId)){throw new ProductNotFoundException("상품을 찾을 수 없습니다");}
         return productRepository.findById(productId).orElseThrow();
     }
 
@@ -35,8 +37,9 @@ public class ProductService implements ProductServiceInterface {
     }
 
     //상품 수정
-    public ProductResponseDto updateProduct(long productId,
+    public ProductResponseDto updateProduct (long productId,
         ProductUpdateRequestDto productUpdateRequestDto) {
+        if(!containsProduct(productId)){throw new ProductNotFoundException("상품을 찾을 수 없습니다");}
         Product product = new Product(
             productId,
             productUpdateRequestDto.name(),
@@ -49,6 +52,7 @@ public class ProductService implements ProductServiceInterface {
 
     //상품 삭제
     public void deleteProduct(long productId) {
+        if(!containsProduct(productId)){throw new ProductNotFoundException("상품을 찾을 수 없습니다");}
         productRepository.delete(productId);
     }
